requirements.txt
aiogram
google-genai
python-dotenv
fastapi
uvicorn
dnspython
